API/LocalInference/requirements.txt
Flask==3.1.0
Flask_Cors==4.0.0
numpy==1.23.5
python-dotenv==1.0.1
torch==2.5.1
transformers==4.46.3
gunicorn==20.1.0
pymongo
